enhance!: Use Uuid newtype to avoid nil Default
Adds an entity::uuid::UuidV4 newtype exported as Uuid that wraps
uuid::Uuid with a different Default trait implementation that uses
Uuid::new_v4() instead of Uuid::nil().

This protects against the possibility of Uuid::default() resulting in a
nil Uuid value that could be inserted into the database.

This also removes the need for custom ActiveModelBehavior::before_save
implementation (which previously was ensuring the value was not nil).

BREAKING CHANGE: The uuid::Uuid type used in entity has been replaced with a
newtype wrapper entity::Uuid that implements only a fraction of the
previous type's behavior.